Crate picket_line

Picket Line

Enable /examples in doc generation: cargo +nightly doc -Z unstable-options -Z rustdoc-scrape-examples=examples

Picket Line is an asynchronous Faktory client. It is heavily informed by faktory-rs, however, it differs significantly in its implementation of the Faktory Work Protocol. In short, faktor-rs is a connection per worker implementation whereas Picket Line is an N workers per connection implementation.

Why Picket Line?

Picket Line is a core library of iDRAMA.cloud, designed based on our decade+ years of experience rapidly developing social media data collection systems (see, our publications for reference). Our “methodology” for crawlers, at a high level, involves viewing the collection process as a set of discrete, asynchronous jobs. As a simple example, imagine that you are building a crawler that will visit a page once a day, and then collect every page linked to from that page. I.e., a one level deep crawl of a site.

While this can be done via a simple loop and a sleep, it could also be done by breaking the problem up into two different sub problems:

Collecting the main page

Collecting all the pages that are linked to from the main page

This can be broken up into jobs as well:

A job that collects the main page

A job that collects a (single) page linked to from the main page

The basic “algorithm” here might be that the main page collection job goes out and retrieves the main page and extracts all the links on it. It then issues a (background) job to collect each of those links. Finally, the main page collection job schedules itself to run again in 24 hours.

These (trivial) jobs are overwhelming I/O bound due to HTTP requests. In our experience, social media data collection systems are overwhelmingly I/O bound (ADD ANY NECESSARY INTUTION HERE). Generally speaking, I/O bound operations is where asynchronous models excel. Thus, Picket Line.

In reality, rapidly developing scalable, continuous social media data collection systems involves building much more complicated work flows than the above example. This niche is what Picket Line is designed to fill.

Why not Picket Line?

Although Picket Line can be run on a multi-threaded async runtime (and the examples use this), there are potential performance pitfalls when doing so. For example, it’s generally a rule of thumb that you should never have too much code in between awaits. This is because asynchronous programming is cooperative; your task will (at a high level) block whatever thread it’s running on until it awaits. You can work around this by launching a separate thread during execution of a Jerb, however you then have to manage signalling via channels etc.

In summary, Picket Line might not be the right choice if your jobs are CPU bound.

Concepts

Picket Line is somewhat higher level than the Faktory protocol itself, and to demarcate these differences, we use a naming convention that differs from the underlying Faktory concept when necessary.

Most users of this library are primarily intersted in implementing different types of Jerbs. A Jerb differs from a [FaktoryJob] in that it is concerned almost entirely with the execution of a job. This includes not just performing the particular work the [FaktoryJob] calls for, but also for handling any additional workflow (e.g., scheduling new [FaktoryJob]s for execution).

Next, we have Werkers, which are mostly similar to Faktory’s Worker concept. A Werker has a set of Jerbs that it knows how to perform. Multiple Werkers can operate off of a single connection the Faktory server, and each of them can listen to their own queues and respond to a unique set of Jerbs.

Structs

JobRoute

PicketLine

PicketLineClient

A basic, high level client that is accessible to crate::Jerbs that are actively executing a job.

Werker

Werkers implement the basic functionality of fetching and executing jobs from a Faktory server.

Traits

Jerb

This is the main trait that should be implemented for the different job types you want to process Of note, is that this trait makes use of the async_trait crate. In turn, this means that all implementations of this trait must be annoted with #[async_trait] as well. It is also cloneable, making use of the dyn-clonable crate