Trait picket_line::Jerb

pub trait Jerb: DynClone {
    fn perform<'life0, 'async_trait>(
        &'life0 self, 
        job: FaktoryJob, 
        picket_line_client: PicketLineClient
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

This is the main trait that should be implemented for the different job types you want to process Of note, is that this trait makes use of the async_trait crate. In turn, this means that all implementations of this trait must be annoted with #[async_trait] as well. It is also cloneable, making use of the dyn-clonable crate

Required methods

fn perform<'life0, 'async_trait>(
&'life0 self,
job: FaktoryJob,
picket_line_client: PicketLineClient
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

This method is what is actually executed when a job is fetched from the Faktory server. An Ok results in the job being ACK‘d to the Faktory server and an Err will result in a `FAIL’ will be send to the Faktory server.